test: fix BaseAcceptanceTest and add acceptance test for reading line list
