+++ src/test/java/nextstep/subway/BaseAcceptanceTest.java
@@ -1,57 +1,30 @@
 package nextstep.subway;
 
-import com.google.common.base.CaseFormat;
-import java.util.List;
-import java.util.stream.Collectors;
-import javax.persistence.Entity;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
+import io.restassured.RestAssured;
+import nextstep.subway.util.DatabaseCleaner;
 import org.junit.jupiter.api.AfterEach;
-import org.springframework.beans.factory.InitializingBean;
+import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.boot.web.server.LocalServerPort;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class BaseAcceptanceTest {
-    public class SpringContainerTest {
-        @Autowired
-        private DatabaseCleaner databaseCleaner;
+    @LocalServerPort
+    int port;
 
-        @AfterEach
-        protected void clearDatabase() {
-            databaseCleaner.tableClear();
+    @Autowired
+    private DatabaseCleaner databaseCleaner;
+
+    @BeforeEach
+    protected void setUp() {
+        if (RestAssured.port == RestAssured.UNDEFINED_PORT) {
+            RestAssured.port = port;
         }
     }
 
-    @Service
-    public class DatabaseCleaner implements InitializingBean {
-        @PersistenceContext
-        private EntityManager entityManager;
-
-        private List<String> tableNames;
-
-        @Override
-        public void afterPropertiesSet() {
-            tableNames = entityManager.getMetamodel().getEntities()
-                    .stream()
-                    .filter(entityType -> entityType.getJavaType().getAnnotation(Entity.class) != null)
-                    .map(entityType -> CaseFormat.UPPER_CAMEL.to(CaseFormat.LOWER_UNDERSCORE, entityType.getName()))
-                    .collect(Collectors.toList());
-        }
-
-        @Transactional
-        public void tableClear() {
-            entityManager.flush();
-            entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY FALSE").executeUpdate();
-            for (String tableName : tableNames) {
-                entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
-                entityManager
-                        .createNativeQuery("ALTER TABLE " + tableName + " ALTER COLUMN ID RESTART WITH 1")
-                        .executeUpdate();
-            }
-            entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY TRUE").executeUpdate();
-        }
+    @AfterEach
+    protected void clearDatabase() {
+        databaseCleaner.clear();
     }
 }
